Manage versions

+++ src/main/java/com/geminicode/hssc/servlet/CheckCardsCron.java
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Locale;
+import java.util.logging.Logger;
 
 /**
  * This servlet save new cards to the Index and the DataStore.
@@ -20,12 +21,20 @@ public class CheckCardsCron extends HttpServlet {
     private final SearchApiService searchApiService = ServiceFactory.get().getSearchApiService();
     private final DatastoreService datastoreService = ServiceFactory.get().getDatastoreService();
 
+    private static final Logger LOGGER = Logger.getLogger(CheckCardsCron.class.getName());
+
+
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        searchApiService.deleteAllCards();
-        datastoreService.removeAllCards();
-        searchApiService.checkNewCards(Locale.FRANCE);
-        searchApiService.checkNewCards(Locale.US);
+        if (searchApiService.isLastedVersion()) {
+            searchApiService.deleteAllCards();
+            datastoreService.removeAllCards();
+            searchApiService.checkNewCards(Locale.FRANCE);
+            searchApiService.checkNewCards(Locale.US);
+        } else {
+            LOGGER.info("Nothing to update");
+        }
+
     }
 }
